requirements.txt
streamlit
pandas
yfinance
requests
beautifulsoup4
newspaper3k
lxml_html_clean
plotly
transformers
torch
torchaudio
torchvision